Python/requirements.txt
fastapi
uvicorn
psycopg2
numpy
scikit-learn
google-genai
